requirements.txt
fastapi==0.110.0
uvicorn
openai>=1.0.0
openai-whisper
python-dotenv
yt-dlp
numpy
pydantic==2.6.1
email-validator
ffmpeg-python
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
